`extract_key.rs` is redundant
Distribute its content between other files.
